docreader/requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
pdfplumber==0.10.3
protobuf==5.27.0
Pillow==10.1.0
PyMuPDF==1.23.8
